test: fix out-of-scope handle in settergetter test
It's not legal to return a v8::Value from a function that has a
v8::HandleScope without wrapping it with v8::HandleScope::Close() first.

Without a call to v8::HandleScope::Close(), the handle gets zapped when
the scope object is destroyed.  That is, in debug builds; in release
builds the behavior is undefined and random.

Take the easy way out by dropping the scope object; the caller already
has one set up.