Ticket #306: Backport thread local storage
This makes thread local storage available to older versions of libuv as
well.  It is essentially a backport of 5d2434bf from the libuv tree, written
by Ben Noordhuis.  But the names were changed, and the functions were turned
into inline functions to avoid having the same non-inline function in
multiple translation units.  The test is a lot more node-based, even if the
things to test more or less follow those tested in libuv.