Expose problem due to This being passed to Unwrap
On Node 0.10.40, the test case will now fail with SIGABRT due to a failed
assertion:

node: ../../nan_object_wrap.h:33:
static T* Nan::ObjectWrap::Unwrap(v8::Local<v8::Object>)
[with T = MyObject]:
Assertion `object->InternalFieldCount() > 0' failed.

This is because This is not an instance of the type in question.
Node >=0.12 seem to report a TypeError: Illegal invocation instead.