AsyncProgressWorker: add template types for .send
This allows for arbitrary types to be sent from .send. To achieve this, the
abstract class can be implemented with a template itself. Either at
implementation or instantiation a type MUST be declared though.

Due to the limitations and / or design of C++ it is not easily possible to
fallback to the old behaviour. In consequence this is a semver major change.
An easy semi-fallback would be to implement the class as

class ProgressWorker : public AsyncProgressWorker<> {}