Add example
Might not be the best example, but this took a lot longer than I had hoped to figure out for myself. Hope sharing it is helpful. If this isn't the best place to put this, would really appreciate an issue opened on the linked project with where it could be surfaced.