AsyncProgressQueueWorker: override workcomplete to ensure WorkProgress gets called
make sure that AsyncProgressQueueWorker::WorkProgress() gets called
before AsyncWorker::WorkComplete(), which otherwise could be risky
without at least 1 ms sleep, so I also added a test for 0 ms sleep.

This might be useful for AsyncProgressWorker as well, but this PR
aims to leave the behavior of AsyncProgressWorker untouched. If we
decide to "fix" AsyncProgressWorker, then we can just move void
WorkProgress() to the AsyncBareProgressWorker

Making sure that WorkProgress() gets called before WorkComplete()
is important to AsyncProgressQueueWorker, since it's aim is to
deliver every event that is generated from progress.Send()

Otherwise, we run the risk that WorkProgress() may get called after
WorkComplete() while this->callback is null and then
this->HandleProgressCallback() wont be called with that last event.
We can sometimes see that happening if we run
test/js/asyncprogressqueueworker-0ms-test.js with this WorkComplete
override removed, or if we run a similar 0ms test instead based on
asyncprogressworker