test/cpp/asyncprogressworker*: delete progress callback in class destructor
Nobody deletes the progress callback in the AsyncProgressWorker tests.

In `test/cpp/asyncprogressworker*.cpp`, the addon functions allocate two
`Nan::Callback` objects, called `progress` and `callback`.

`callback` is passed into the `AsyncWorker` constructor, and gets
deleted by the `AsyncWorker` destructor after work is complete.

`progress` is passed into the user-level test class.  It is called
in `HandleProgressCallback()` but it never gets deleted.

I believe this is an error, and it's quite a big one, since developers look
at these tests for examples of how to properly implement these workers.