AsyncProgressQueueWorker doesnt support ExecutionProgress.Signal()
`AsyncProgressQueueWorker::ExecutionProgress.Signal()` is a no-op
if we leave things as-is.  We can try to fix that, but there isn't
any point.  If somebody wants that functionality, they should just
use an `AsyncProgressWorkerBase<T>`  Let's just remove the entry
point by swapping out the `AsyncBareProgressWorker<T>` with a
specialized `AsyncBareProgressQueueWorker<T>` that simply doesn't
have an `ExecutionProgress.Signal()`