remove test/js/asyncprogressworker-1ms-test.js
This test usually passes, but it's a little flakey and fails once
in a blue moon.  `AsyncProgressWorker` isn't great at what it does
if you don't wait enough time between calls. That's why I created
`AsyncProgressQueueWorker`, which can do fine without any sleeps.

The intention of this test was to show that `AsyncProgressQueueWorker`
handles things better without artificial sleeps. It's main purpose
was simply to demonstrate this in the original pull request.

We don't really need this particular test to stay in the
_NAN_ repository. Since it fails sometimes, let's kill it.