test/*/asyncprogressqueueworker*: kill sleeps & rename `a` to `doProgress`
As described in #703 , these varying sleep tests for the
`AsyncProgressQueueWorker` were created purely for the sake of the
original pull request.

These tests are somewhat redundant now as a result of the following commit:

a5db7f097ef78a68622889f9efbee32ae495c437 -
AsyncProgressQueueWorker: override workcomplete to ensure WorkProgress gets called

So now we can remove any remnants of `Sleep` from the cpp addon, and
go back to a single js test for the basic `AsyncProgressQueueWorker`
without any forced sleeps.  This also serves as a better coding example
for the `AsyncProgressQueueWorker` class template.

...and while we're here, rename the addon entry point `a` to `doProgress`