avoid access after free in returnvalue test
The isolate can be Disposed due to the test finishing before the destructor
of the static Global fires. Resetting said Global avoids this possibility.
The point of the test is to verify that returning a Global is possible, not
that this interaction of Global's destructor and Node shutting down is safe.